refactor: move common code to TimeTrait
